handle errors in multiple queries correctly
This is kinda hackish

flush-pending-data tries to find the ending EOF/OK packet for the last
packet. In case of multiple queries and one of the quries could produce
error, and flush-pending-data would break in this case, because it
skipped from the first error detection, and it doesn't end with an
EOF/OK packet, therefore, flush-pending-data will keep polling the
socket until network error.

There is still a bug when multiple queries are sent in a line,
it only returns the first result set, and flush-pending-data will only
be called when the next send-sql is called.