Initial handshak packet could be an ERR packet
from https://dev.mysql.com/doc/internals/en/connection-phase.html:

It starts with the client connect()ing to the server which may send
a ERR packet and finish the handshake or send a Initial Handshake
Packet which the client answers with a Handshake Response Packet. At
this stage client can request SSL connection, in which case an SSL
communication channel is established before client sends its
authentication response.

This is an example error message:

Host '192.168.xx.xx' is blocked because of many connection errors;
unblock with 'mysqladmin flush-hosts'