Do not alter auth data
auth data is sent as a string in the protocol, and read-string converts
it from binary! to string!, which seems reasonable. However, converting
to string! from binary! in R3 also does line break conversion, which
means #{0D} will be converted to "^A", instead of "^M". So, if the auth
data contains #{0D}, the auth data used in the authentication
calculation would be wrong, and cause a "access denied" error. Because
the auth data is a random string generated by the server, we might not
see this all the time, but occasionally.