Fix sending multiple queries in one call
pl/more-results? was not properly checked, which caused the driver to
permaturely return an empty result, and use it as the result of next
query (query and result become out-of-sync)

Also separate state into two variables: one for the packet reading
(lower level) and another one is for the mysql protocol state. So it
doesn't need to switch state to 'reading-packet-header when it's in
middle of processing a response and use 'saved-status to remember what
state it was in. This also makes handling chunked packets easier, as no
mysql state changes when a chunk of a big packet is received. This
pretty much makes packet-state local to tcp-awake.